refactor: Atomicデザイン化 テンプレートとページの分離
