+++ src/components/pages/Bbs.vue
@@ -1,43 +1,46 @@
 <template>
-  <div class="bbs">
-    <h1>一言掲示板</h1>
-    <p>
-      足跡や、あいさつ、キリ番報告、等に使用していただいて構いません。
-      <br>
-      誹謗中傷、個人情報などの書き込みはご遠慮ください。
-    </p>
-    <div class="message-form">
-      <div class="message-form-group">
-        <label for="nameInput">HN(ハンドルネーム)</label>
-        <input type="text" id="nameInput" maxlength="20" v-model="name">
+  <default-template>
+    <div class="bbs">
+      <h1>一言掲示板</h1>
+      <p>
+        足跡や、あいさつ、キリ番報告、等に使用していただいて構いません。
+        <br>
+        誹謗中傷、個人情報などの書き込みはご遠慮ください。
+      </p>
+      <div class="message-form">
+        <div class="message-form-group">
+          <label for="nameInput">HN(ハンドルネーム)</label>
+          <input type="text" id="nameInput" maxlength="20" v-model="name">
+        </div>
+        <div class="message-form-group">
+          <label for="messageInput">メッセージ</label>
+          <input type="text" id="messageInput" maxlength="100" v-model="message">
+        </div>
+        <button type="button" @click="sendMessage">送信</button>
       </div>
-      <div class="message-form-group">
-        <label for="messageInput">メッセージ</label>
-        <input type="text" id="messageInput" maxlength="100" v-model="message">
+      <div>
+        <ul class="message-list">
+          <li v-for="(item, key) in messageList" v-bind:key="key">
+            <p>
+              {{item.message}} by {{item.name}}
+            </p>
+            <p class="time-label">
+              {{item.createdAt}}
+            </p>
+          </li>
+        </ul>
       </div>
-      <button type="button" @click="sendMessage">送信</button>
     </div>
-    <div>
-      <ul class="message-list">
-        <li v-for="(item, key) in messageList" v-bind:key="key">
-          <p>
-            {{item.message}} by {{item.name}}
-          </p>
-          <p class="time-label">
-            {{item.createdAt}}
-          </p>
-        </li>
-      </ul>
-    </div>
-  </div>
+  </default-template>
 </template>
 
 <script lang="ts">
 import moment from 'moment';
 import Vue from 'vue';
 import Component from 'vue-class-component';
+import DefaultTemplate from '@/components/templates/Default';
 import firebase from 'firebase';
-import database from '../firebase-config';
+import database from '@/firebase-config';
 
 interface Message {
   name: string;
@@ -52,7 +55,11 @@ interface BbsData {
   message: string;
 }
 
-@Component
+@Component({
+  components: {
+    DefaultTemplate,
+  },
+})
 export default class Bbs extends Vue implements BbsData {
   messageList: Message[] = [];
   name: string = '';
