+++ src/router/index.ts
@@ -1,10 +1,10 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Top from '@/components/Top.vue';
-import Disco from '@/components/Disco.vue';
-import Bbs from '@/components/Bbs.vue';
-import Member from '@/components/Member.vue';
-import NotFound from '@/components/NotFound.vue';
+import Top from '@/components/pages/Top.vue';
+import Disco from '@/components/pages/Disco.vue';
+import Bbs from '@/components/pages/Bbs.vue';
+import Member from '@/components/pages/Member.vue';
+import NotFound from '@/components/pages/NotFound.vue';
 
 Vue.use(Router);
